# server/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Home API
  description: |
    API for controlling smart home devices including Philips Hue lights and Nanoleaf panels.
    
    ## Overview
    This API provides a unified interface for managing multiple smart lighting systems:
    - **Philips Hue** - Bridge-based smart lights with rooms/groups support
    - **Nanoleaf** - LED panels with effects support
    
    ## Authentication
    Device authentication is handled through pairing processes:
    - **Hue**: Press the link button on the bridge, then call `/api/hue/authenticate`
    - **Nanoleaf**: Hold power button for 5-7 seconds until lights flash, then call `/api/nanoleaf/authenticate`
    
    ## Color Modes
    Devices support multiple color modes:
    - `hs` - Hue/Saturation mode
    - `ct` - Color temperature mode (Kelvin)
    - `effect` - Device-specific effects
  version: 1.0.0
  contact:
    name: Smart Home API
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /api/devices/{id}:
    put:
      tags:
        - General
      summary: Update device state
      description: Update any device by ID (works for both Hue and Nanoleaf)
      operationId: updateDevice
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceStateUpdate'
      responses:
        '200':
          description: Device updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Room Endpoints
  /api/rooms:
    get:
      tags:
        - Rooms
      summary: Get all rooms
      description: Returns all rooms with their devices and groups
      operationId: getRooms
      responses:
        '200':
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
    post:
      tags:
        - Rooms
      summary: Create a room
      description: Create a new room for organizing devices
      operationId: createRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Room name
                  example: "Kitchen"
      responses:
        '201':
          description: Room created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
...
  /api/groups:
    get:
      tags:
        - Groups
      summary: Get all device groups
      description: Returns all device groups with their devices
      operationId: getGroups
      responses:
        '200':
          description: List of device groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceGroup'
    post:
      tags:
        - Groups
      summary: Create a device group
      description: Create a new device group for controlling multiple devices together
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Group name
                  example: "Kitchen Table Lights"
                roomId:
                  type: string
                  format: uuid
                  nullable: true
                  description: Optional room ID to associate group with
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceGroup'
  /api/groups/{id}:
    get:
      tags:
        - Groups
      summary: Get a device group
      description: Returns a specific device group by ID
      operationId: getGroup
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Group details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceGroup'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Groups
      summary: Update a device group
      description: Update a group's name or room association
      operationId: updateGroup
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                roomId:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        '200':
          description: Group updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceGroup'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Groups
      summary: Delete a device group
      description: Delete a device group (devices remain but are no longer grouped)
      operationId: deleteGroup
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '204':
          description: Group deleted successfully
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/groups/{groupId}/devices:
    post:
      tags:
        - Groups
      summary: Add device to group
      description: Add a device to a device group
      operationId: addDeviceToGroup
      parameters:
        - name: groupId
          in: path
          required: true
          description: Group ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - deviceExternalId
              properties:
                deviceExternalId:
                  type: string
                  description: Device external ID (e.g., hue-1)
      responses:
        '201':
          description: Device added to group successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Group or device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/groups/{groupId}/devices/{deviceExternalId}:
    delete:
      tags:
        - Groups
      summary: Remove device from group
      description: Remove a device from a device group
      operationId: removeDeviceFromGroup
      parameters:
        - name: groupId
          in: path
          required: true
          description: Group ID
          schema:
            type: string
            format: uuid
        - name: deviceExternalId
          in: path
          required: true
          description: Device external ID (e.g., hue-1)
          schema:
            type: string
      responses:
        '204':
          description: Device removed from group successfully
        '404':
          description: Group or device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/groups/{id}/state:
    put:
      tags:
        - Groups
      summary: Control all devices in group
      description: Set the state of all devices in a group
      operationId: setGroupState
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceStateUpdate'
      responses:
        '200':
          description: Group state updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        deviceId:
                          type: string
                        success:
                          type: boolean
                        error:
                          type: string
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Hue Endpoints
...
  /api/hue/groups/{id}:
    put:
      tags:
        - Hue
      summary: Update Hue group state
      description: Update the state of all lights in a group (room)
      operationId: updateHueGroup
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID (e.g., hue-group-1)
          schema:
            type: string
            example: hue-group-1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceStateUpdate'
      responses:
        '200':
          description: Group updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Nanoleaf Endpoints
...
components:
  parameters:
    DeviceId:
      name: id
      in: path
      required: true
      description: Device ID (e.g., hue-1 for Hue lights or UUID for Nanoleaf)
      schema:
        type: string
        example: hue-1
...
    GroupId:
      name: id
      in: path
      required: true
      description: Device group UUID
      schema:
        type: string
        format: uuid
        example: 550e8400-e29b-41d4-a716-446655440002
  schemas:
    Light:
      type: object
      description: Unified light/device representation
      properties:
        id:
          type: string
          description: Unique device identifier
          example: hue-1
        name:
          type: string
          description: User-friendly device name
          example: Living Room Lamp
        type:
          type: string
          enum: [hue, nanoleaf]
          description: Device type/manufacturer
        state:
          $ref: '#/components/schemas/DeviceState'
        reachable:
          type: boolean
          description: Whether the device is reachable on the network
        model:
          type: string
          description: Device model identifier
          example: LCT007
        capabilities:
          type: object
          properties:
            color:
              type: boolean
              description: Whether device supports color
            colorTemp:
              type: boolean
              description: Whether device supports color temperature
        roomId:
          type: string
          format: uuid
          nullable: true
          description: Room ID the device is assigned to
        roomName:
          type: string
          nullable: true
          description: Name of the room the device is assigned to
        groups:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
          description: Groups the device belongs to
    Room:
      type: object
      description: Room for organizing devices
      properties:
        id:
          type: string
          format: uuid
          description: Unique room identifier
        name:
          type: string
          description: Room name
          example: Kitchen
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        devices:
          type: array
          items:
            $ref: '#/components/schemas/Light'
          description: Devices in this room
        groups:
          type: array
          items:
            $ref: '#/components/schemas/DeviceGroup'
          description: Device groups in this room
    DeviceGroup:
      type: object
      description: Group of devices for simultaneous control
      properties:
        id:
          type: string
          format: uuid
          description: Unique group identifier
        name:
          type: string
          description: Group name
          example: Kitchen Table Lights
        roomId:
          type: string
          format: uuid
          nullable: true
          description: Room this group belongs to
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        room:
          $ref: '#/components/schemas/Room'
          nullable: true
          description: Room this group belongs to
        devices:
          type: array
          items:
            type: object
            properties:
              device:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  externalId:
                    type: string
                  name:
                    type: string
                  type:
                    type: string
          description: Devices in this group
    DeviceState:
      type: object
      description: Current state of a device
      properties:
        on:
          type: boolean
          description: Whether the device is on
        brightness:
          type: integer
          minimum: 0
          maximum: 100
          description: Brightness level (0-100)
        color:
          $ref: '#/components/schemas/ColorHSB'
        colorTemp:
          type: integer
          minimum: 2000
          maximum: 6500
          description: Color temperature in Kelvin
          example: 4000
        colorMode:
          type: string
          enum: [hs, ct, effect]
          description: Current color mode
    DeviceStateUpdate:
      type: object
      description: Partial device state for updates
      properties:
        on:
          type: boolean
          description: Turn device on or off
        brightness:
          type: integer
          minimum: 0
          maximum: 100
          description: Brightness level (0-100)
        color:
          $ref: '#/components/schemas/ColorHSB'
        colorTemp:
          type: integer
          minimum: 2000
          maximum: 6500
          description: Color temperature in Kelvin
    ColorHSB:
      type: object
      description: HSB color representation
      properties:
        hue:
          type: number
          minimum: 0
          maximum: 360
          description: Hue value (0-360 degrees)
          example: 240
        saturation:
          type: number
          minimum: 0
          maximum: 100
          description: Saturation percentage (0-100)
          example: 100
        brightness:
          type: number
          minimum: 0
          maximum: 100
          description: Brightness percentage (0-100)
          example: 75
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: Device not found